+++ mfa/Common.py
@@ -1,4 +1,8 @@
+import datetime
+from random import randint
+
 from django.conf import settings
+from django.contrib.auth import get_user_model
 from django.core.mail import EmailMessage
 
 try:
@@ -24,3 +28,20 @@ def get_redirect_url():
         ),
         "reg_success_msg": getattr(settings, "MFA_SUCCESS_REGISTRATION_MSG"),
     }
+
+
+def get_username_field():
+    User = get_user_model()
+    USERNAME_FIELD = getattr(User, "USERNAME_FIELD", "username")
+    return User, USERNAME_FIELD
+
+
+def set_next_recheck():
+    if getattr(settings, "MFA_RECHECK", False):
+        delta = datetime.timedelta(
+            seconds=randint(settings.MFA_RECHECK_MIN, settings.MFA_RECHECK_MAX)
+        )
+        return {
+            "next_check": datetime.datetime.timestamp(datetime.datetime.now() + delta)
+        }
+    return {}